refactor ArtistGrid components to separate files
